handle base-path and filename args
