+++ node/src/command.rs
@@ -10,17 +10,13 @@ use frame_benchmarking_cli::{BenchmarkCmd, ExtrinsicFactory, SUBSTRATE_REFERENCE
 use quantus_runtime::{Block, EXISTENTIAL_DEPOSIT};
 use rusty_crystals_hdwallet::wormhole::WormholePair;
 use rusty_crystals_hdwallet::{generate_mnemonic, HDLattice};
-use sc_cli::{Error, KeySubcommand, SubstrateCli};
+use sc_cli::{SubstrateCli};
 use sc_network::config::{NodeKeyConfig, Secret};
-use sc_network::{Keypair};
-use sc_service::{BasePath, BlocksPruning, PartialComponents, PruningMode};
+use sc_service::{BlocksPruning, PartialComponents, PruningMode};
 use sp_core::crypto::AccountId32;
 use sp_core::crypto::Ss58Codec;
 use sp_keyring::Sr25519Keyring;
 use sp_runtime::traits::IdentifyAccount;
-use std::io::Write;
-use std::path::PathBuf;
-use std::{fs, io};
 #[derive(Debug, PartialEq)]
 pub struct QuantusKeyDetails {
     pub address: String,
@@ -29,41 +25,6 @@ pub struct QuantusKeyDetails {
     pub secret_key_hex: String, // Secret key, hex encoded with "0x" prefix
     pub seed_hex: String,       // Derived seed, hex encoded with "0x" prefix
     pub secret_phrase: Option<String>, // Mnemonic phrase
-}
-
-const NODE_KEY_DILITHIUM_FILE: &str = "secret_dilithium";
-const DEFAULT_NETWORK_CONFIG_PATH: &str = "network";
-
-/// Returns the value of `base_path` or the default_path if it is None
-pub(crate) fn base_path_or_default(
-    base_path: Option<BasePath>,
-    executable_name: &String,
-) -> BasePath {
-    base_path.unwrap_or_else(|| BasePath::from_project("", "", executable_name))
-}
-
-/// Returns the default path for configuration  directory based on the chain_spec
-pub(crate) fn build_config_dir(base_path: &BasePath, chain_spec_id: &str) -> PathBuf {
-    base_path.config_dir(chain_spec_id)
-}
-
-/// Returns the default path for the network configuration inside the configuration dir
-pub(crate) fn build_net_config_dir(config_dir: &PathBuf) -> PathBuf {
-    config_dir.join(DEFAULT_NETWORK_CONFIG_PATH)
-}
-
-/// Returns the default path for the network directory starting from the provided base_path
-/// or from the default base_path.
-pub(crate) fn build_network_key_dir_or_default(
-    base_path: Option<BasePath>,
-    chain_spec_id: &str,
-    executable_name: &String,
-) -> PathBuf {
-    let config_dir = build_config_dir(
-        &base_path_or_default(base_path, executable_name),
-        chain_spec_id,
-    );
-    build_net_config_dir(&config_dir)
 }
 
 pub fn generate_quantus_key(
@@ -155,55 +116,6 @@ pub fn generate_quantus_key(
     }
 }
 
-/// This is copied from sc-cli and adapted to dilithium
-fn generate_key_in_file(
-    file: &Option<PathBuf>,
-    chain_spec_id: Option<&str>,
-    base_path: &Option<PathBuf>,
-    default_base_path: bool,
-    executable_name: Option<&String>,
-    keypair: Option<&Keypair>,
-) -> Result<(), Error> {
-    let kp: Keypair;
-    if let Some(k) = keypair {
-        kp = k.clone();
-    } else {
-        kp = Keypair::generate_dilithium();
-    }
-    let file_data = kp.to_protobuf_encoding().unwrap();
-
-    match (file, base_path, default_base_path) {
-        (Some(file), None, false) => fs::write(file, file_data)?,
-        (None, Some(_), false) | (None, None, true) => {
-            let network_path = build_network_key_dir_or_default(
-                base_path.clone().map(BasePath::new),
-                chain_spec_id.unwrap_or_default(),
-                executable_name.ok_or(Error::Input("Executable name not provided".into()))?,
-            );
-
-            fs::create_dir_all(network_path.as_path())?;
-
-            let key_path = network_path.join(NODE_KEY_DILITHIUM_FILE);
-            if key_path.exists() {
-                eprintln!("Skip generation, a key already exists in {:?}", key_path);
-                return Err(Error::KeyAlreadyExistsInPath(key_path));
-            } else {
-                eprintln!("Generating key in {:?}", key_path);
-                fs::write(key_path, file_data)?
-            }
-        }
-        (None, None, false) => io::stdout().lock().write_all(&file_data)?,
-        (_, _, _) => {
-            // This should not happen, arguments are marked as mutually exclusive.
-            return Err(Error::Input("Mutually exclusive arguments provided".into()));
-        }
-    }
-
-    eprintln!("{}", kp.public().to_peer_id());
-
-    Ok(())
-}
-
 impl SubstrateCli for Cli {
     fn impl_name() -> String {
         "Quantus Node".into()
@@ -256,20 +168,7 @@ pub fn run() -> sc_cli::Result<()> {
     match &cli.subcommand {
         Some(Subcommand::Key(cmd)) => {
             match cmd {
-                QuantusKeySubcommand::Sc(sc_cmd) => match sc_cmd {
-                    KeySubcommand::GenerateNodeKey(gen_cmd) => {
-                        let chain_spec = cli.load_spec(gen_cmd.chain.as_deref().unwrap_or(""))?;
-                        generate_key_in_file(
-                            &None,
-                            Some(chain_spec.id()),
-                            &None,
-                            true,
-                            Some(&Cli::executable_name()),
-                            None,
-                        )
-                    }
-                    _ => sc_cmd.run(&cli),
-                },
+                QuantusKeySubcommand::Sc(sc_cmd) => sc_cmd.run(&cli),
                 QuantusKeySubcommand::Quantus {
                     scheme,
                     seed,
@@ -524,7 +423,6 @@ mod tests {
         EXPECTED_PUBLIC_KEY_HEX, EXPECTED_SECRET_KEY_HEX, TEST_ADDRESS, TEST_MNEMONIC,
         TEST_SEED_HEX,
     };
-    use tempfile::TempDir;
 
     #[test]
     fn test_generate_quantus_key_standard_new_mnemonic() {
@@ -648,112 +546,4 @@ mod tests {
         assert_eq!(details.secret_key_hex, expected_secret_key_hex);
     }
 
-    #[test]
-    fn test_generate_key_in_file_explicit_path() {
-        // Setup: Create a temporary directory and file path.
-        let temp_dir = TempDir::new().unwrap();
-        let file_path = temp_dir.path().join("test.key");
-        let original_keypair = Keypair::generate_dilithium();
-        // Generate and write keypair.
-        let result = generate_key_in_file(
-            &Some(file_path.clone()),
-            None,
-            &None,
-            false,
-            Some(&"test-exec".to_string()),
-            Some(&original_keypair),
-        );
-        assert!(result.is_ok(), "Failed to generate key: {:?}", result);
-
-        // Read the file contents.
-        let file_data = fs::read(&file_path).expect("Failed to read file");
-
-        // Deserialize the keypair.
-        let deserialized_keypair =
-            Keypair::from_protobuf_encoding(&file_data).expect("Failed to deserialize keypair");
-
-        // Verify the deserialized keypair matches the original.
-        assert_eq!(
-            deserialized_keypair.public().to_peer_id(),
-            original_keypair.public().to_peer_id(),
-            "Deserialized public key does not match original"
-        );
-    }
-
-    #[test]
-    fn test_generate_key_in_file_default_path() {
-        // Setup: Create a temporary directory as base path.
-        let temp_dir = TempDir::new().unwrap();
-        let base_path = Some(temp_dir.path().to_path_buf());
-        let chain_spec_id = "test-chain";
-        let executable_name = "test-exec";
-        let original_keypair = Keypair::generate_dilithium();
-
-        // Generate and write keypair.
-        let result = generate_key_in_file(
-            &None,
-            Some(chain_spec_id),
-            &base_path,
-            false,
-            Some(&executable_name.to_string()),
-            Some(&original_keypair),
-        );
-        assert!(result.is_ok(), "Failed to generate key: {:?}", result);
-
-        // Construct the expected key file path.
-        let expected_path = temp_dir
-            .path()
-            .join("chains")
-            .join(chain_spec_id)
-            .join("network")
-            .join(NODE_KEY_DILITHIUM_FILE);
-
-        // Read the file contents.
-        let file_data = fs::read(&expected_path).expect("Failed to read file");
-
-        // Deserialize the keypair.
-        let deserialized_keypair =
-            Keypair::from_protobuf_encoding(&file_data).expect("Failed to deserialize keypair");
-
-        // Verify the deserialized keypair matches the original.
-        assert_eq!(
-            deserialized_keypair.public().to_peer_id(),
-            original_keypair.public().to_peer_id(),
-            "Deserialized public key does not match original"
-        );
-    }
-
-    #[test]
-    fn test_generate_key_in_file_key_already_exists() {
-        // Setup: Create a temporary directory and pre-create the key file.
-        let temp_dir = TempDir::new().unwrap();
-        let base_path = Some(temp_dir.path().to_path_buf());
-        let chain_spec_id = "test-chain";
-        let executable_name = "test-exec";
-
-        // Create the network directory and key file.
-        let network_path = temp_dir
-            .path()
-            .join("chains")
-            .join(chain_spec_id)
-            .join("network");
-        fs::create_dir_all(&network_path).unwrap();
-        let key_path = network_path.join(NODE_KEY_DILITHIUM_FILE);
-        fs::write(&key_path, vec![0u8; 8]).unwrap(); // Write dummy data.
-
-        // Attempt to generate key (should fail due to existing file).
-        let result = generate_key_in_file(
-            &None,
-            Some(chain_spec_id),
-            &base_path,
-            false,
-            Some(&executable_name.to_string()),
-            None,
-        );
-        assert!(
-            matches!(result, Err(Error::KeyAlreadyExistsInPath(_))),
-            "Expected KeyAlreadyExistsInPath error, got: {:?}",
-            result
-        );
-    }
 }
